+++ lua/error-lens/highlight.lua
@@ -1,14 +1,11 @@
-local api = vim.api
-
 local config = require("error-lens.config")
 
 local function set_highlight(buf, diagnostic, fg, bg)
-	api.nvim_buf_add_highlight(buf, config.namespace, fg, diagnostic.lnum, 0, -1)
+	vim.api.nvim_buf_add_highlight(buf, config.namespace, fg, diagnostic.lnum, 0, -1)
 	vim.api.nvim_buf_set_extmark(buf, config.namespace, diagnostic.lnum, 0, {
 		hl_mode = "combine",
 		hl_eol = true,
 		hl_group = bg,
-		end_row = diagnostic.lnum + 1,
 		virt_text = {
 			{ string.rep(" ", 5) .. diagnostic.message, fg },
 		},
@@ -17,7 +14,7 @@ local function set_highlight(buf, diagnostic, fg, bg)
 end
 
 local function clear_red_highlights(buf)
-	api.nvim_buf_clear_namespace(buf, config.namespace, 0, -1)
+	vim.api.nvim_buf_clear_namespace(buf, config.namespace, 0, -1)
 end
 
 local function update_highlights(buf, diagnostics)
@@ -38,4 +35,5 @@ end
 
 return {
 	update_highlights = update_highlights,
+	clear_red_highlights = clear_red_highlights,
 }